requirements.txt
requests
fastapi
pandas
uvicorn
jinja2
numpy
python-multipart
torch
transformers
qdrant-client
ollama
openai
python-dotenv
pyjwt
passlib[bcrypt]
bcrypt==4.0.1
pika
celery